refactor(core): change SidebarUserSettings to accept a list of providerSettings
This allows to configure custom provider settings without the need to modify the core package.
