+++ packages/core/src/layout/Sidebar/UserSettings.tsx
@@ -14,36 +14,22 @@
  * limitations under the License.
  */
 
-import {
-  githubAuthApiRef,
-  gitlabAuthApiRef,
-  googleAuthApiRef,
-  identityApiRef,
-  oauth2ApiRef,
-  oktaAuthApiRef,
-  microsoftAuthApiRef,
-  useApi,
-  configApiRef,
-} from '@backstage/core-api';
+import { identityApiRef, useApi } from '@backstage/core-api';
 import Collapse from '@material-ui/core/Collapse';
 import SignOutIcon from '@material-ui/icons/MeetingRoom';
-import Star from '@material-ui/icons/Star';
 import React, { useContext, useEffect } from 'react';
 import { SidebarContext } from './config';
 import { SidebarItem } from './Items';
-import {
-  OAuthProviderSettings,
-  OIDCProviderSettings,
-  UserProfile as SidebarUserProfile,
-} from './Settings';
+import { UserProfile as SidebarUserProfile } from './Settings';
 
-export function SidebarUserSettings() {
+type SidebarUserSettingsProps = { providerSettings?: React.ReactNode };
+
+export function SidebarUserSettings({
+  providerSettings,
+}: SidebarUserSettingsProps) {
   const { isOpen: sidebarOpen } = useContext(SidebarContext);
   const [open, setOpen] = React.useState(false);
   const identityApi = useApi(identityApiRef);
-  const configApi = useApi(configApiRef);
-  const providersConfig = configApi.getOptionalConfig('auth.providers');
-  const providers = providersConfig?.keys() ?? [];
 
   // Close the provider list when sidebar collapse
   useEffect(() => {
@@ -54,48 +40,8 @@ export function SidebarUserSettings() {
     <>
       <SidebarUserProfile open={open} setOpen={setOpen} />
       <Collapse in={open} timeout="auto">
-        {providers.includes('google') && (
-          <OIDCProviderSettings
-            title="Google"
-            apiRef={googleAuthApiRef}
-            icon={Star}
-          />
-        )}
-        {providers.includes('microsoft') && (
-          <OIDCProviderSettings
-            title="Microsoft"
-            apiRef={microsoftAuthApiRef}
-            icon={Star}
-          />
-        )}
-        {providers.includes('github') && (
-          <OAuthProviderSettings
-            title="GitHub"
-            apiRef={githubAuthApiRef}
-            icon={Star}
-          />
-        )}
-        {providers.includes('gitlab') && (
-          <OAuthProviderSettings
-            title="GitLab"
-            apiRef={gitlabAuthApiRef}
-            icon={Star}
-          />
-        )}
-        {providers.includes('okta') && (
-          <OIDCProviderSettings
-            title="Okta"
-            apiRef={oktaAuthApiRef}
-            icon={Star}
-          />
-        )}
-        {providers.includes('oauth2') && (
-          <OIDCProviderSettings
-            title="YourOrg"
-            apiRef={oauth2ApiRef}
-            icon={Star}
-          />
-        )}
+        {providerSettings}
+
         <SidebarItem
           icon={SignOutIcon}
           text="Sign Out"
